Refactor the heterogenous kernel event queue and process task queue.
The original design was nearly a (limited) direct map to BSD's kqueue.
In order to make epoll integration easier, create a minor abstraction
and isolate most of the kqueue specific functions to kqueue.c and kqueue.h.

When epoll is introduced, there will be corresponding epoll.c and epoll.h
sources providing the same interface specified by kernelq.h.

This revision tightly couples the task queue and kernel queue where
scheduled operations are directly associated with the target tasks to
be enqueued for subsequent execution.

The interface for event subscription has been consolidated with the creation
of the Event type. This union data structure provides the necessary information
for identifying the event with the kernelq abstraction and is joined with the task
on Link instances to form a dispatchable operation.

- Implement level-triggered I/O events(system.io is strictly edge triggered).
- Implement scheduler meta events allowing actuation and termination to be observed.
- Implement "never" events for invariants and miscellaneous utility.